requirements.txt
numpy==1.18.2
tqdm==4.46.1
gensim==3.8.3
fire==0.2.1
python_telegram_bot==12.5.1
scipy==1.4.1
spacy==2.2.4
pandas==1.0.4
sklearn_crfsuite==0.3.6
PyYAML==5.3.1
scikit_learn==0.23.1
telegram==0.0.1
